+++ .idea/runConfigurations/Main.xml
@@ -2,7 +2,7 @@
   <configuration default="false" name="Main" type="Application" factoryName="Application">
     <option name="ALTERNATIVE_JRE_PATH" value="1.8" />
     <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="true" />
-    <option name="MAIN_CLASS_NAME" value="projlab.Main" />
+    <option name="MAIN_CLASS_NAME" value="projlabController.Main" />
     <module name="TravisGradleProjlab_main" />
     <option name="WORKING_DIRECTORY" value="file://$PROJECT_DIR$" />
     <method />
