+++ src/main/java/projabModel/CyclicIterator.java
@@ -1,4 +1,4 @@
-package projlab;
+package projabModel;
 
 import java.util.Collection;
 import java.util.Iterator;
