+++ src/main/java/projabModel/Direction.java
@@ -1,4 +1,4 @@
-package projlab;
+package projabModel;
 
 
 /**
