+++ src/main/java/projabModel/Field.java
@@ -1,4 +1,4 @@
-package projlab;
+package projabModel;
 
 import java.util.HashMap;
