+++ src/main/java/projabModel/Game.java
@@ -1,9 +1,6 @@
-package projlab;
+package projabModel;
 
-import java.io.File;
 import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 //TODO: Add Javadoc
 public class Game {
