+++ src/main/java/projabModel/GameElement.java
@@ -1,4 +1,4 @@
-package projlab;
+package projabModel;
 
 /**
  * <h1>GameElement</h1>
