+++ src/main/java/projabModel/IMovable.java
@@ -1,4 +1,4 @@
-package projlab;
+package projabModel;
 
 public interface IMovable {
     boolean move(Direction direction, double power);
