+++ src/main/java/projabModel/Player.java
@@ -1,4 +1,4 @@
-package projlab;
+package projabModel;
 
 import java.util.Random;
