+++ src/main/java/projabModel/Switch.java
@@ -1,4 +1,4 @@
-package projlab;
+package projabModel;
 
 /**
  * <h1>Switch</h1>
